Feature : Yolo v8 Seg Integration

+++ hot_fair_utilities/preprocessing/preprocess.py
@@ -4,9 +4,7 @@ import os
 from ..georeferencing import georeference
 from .clip_labels import clip_labels
 from .fix_labels import fix_labels
-from .multimasks_from_polygons import multimasks_from_polygons
 from .reproject_labels import reproject_labels_to_epsg3857
-from .multimasks_from_polygons import multimasks_from_polygons
 
 
 def preprocess(
@@ -18,6 +16,7 @@ def preprocess(
     multimasks=False,
     input_contact_spacing=8,  # only required if multimasks is set to true
     input_boundary_width=3,  # only required if mulltimasks is set to true
+    epsg=3857,
 ) -> None:
     """Fully preprocess the input data.
 
@@ -63,6 +62,7 @@ def preprocess(
         )
     """
     # Check if rasterizing options are valid
+    assert epsg in (4326, 3857), "Projection not supported"
     if rasterize:
         assert (
             rasterize_options is not None
@@ -80,24 +80,37 @@ def preprocess(
     os.makedirs(output_path, exist_ok=True)
 
     if georeference_images:
-        georeference(input_path, f"{output_path}/chips")
+        georeference(input_path, f"{output_path}/chips", epsg=epsg)
 
     fix_labels(
         f"{input_path}/labels.geojson",
         f"{output_path}/corrected_labels.geojson",
     )
+    if epsg == 3857:
+        reproject_labels_to_epsg3857(
+            f"{output_path}/corrected_labels.geojson",
+            f"{output_path}/labels_epsg3857.geojson",
+        )
 
-    reproject_labels_to_epsg3857(
-        f"{output_path}/corrected_labels.geojson",
-        f"{output_path}/labels_epsg3857.geojson",
+    clip_labels(
+        input_path,
+        output_path,
+        rasterize,
+        rasterize_options,
+        all_geojson_file=(
+            f"{output_path}/corrected_labels.geojson"
+            if epsg == 4326
+            else f"{output_path}/labels_epsg3857.geojson"
+        ),
+        epsg=epsg,
     )
 
-    clip_labels(input_path, output_path, rasterize, rasterize_options)
-
     os.remove(f"{output_path}/corrected_labels.geojson")
-    os.remove(f"{output_path}/labels_epsg3857.geojson")
+    if epsg == 3857:
+        os.remove(f"{output_path}/labels_epsg3857.geojson")
 
     if multimasks:
+        from .multimasks_from_polygons import multimasks_from_polygons
 
         assert os.path.isdir(
             f"{output_path}/chips"
